Preset creation now works in back-end.
